Добавлен функционал манипуляций с ВАП при редактировании UIS-инжектора, улучшен для CR-инжектора. In progress 01.10.2020 13.30
- в InjectorUisVAP @NamedEntityGraph(name = "UisInjectorsByVAP"....) и getInjectors() для получения списка инжекторов, к которым привязан этот ВАП
- в VapService добавлен метод  VAP findByVapName(String vapName), его реализует UisVapService
- в InjectorUisVapRepository добавлен метод InjectorUisVAP findByProfileName(String vapName) с @EntityGraph(value = "UisInjectorsByVAP", type = EntityGraph.EntityGraphType.LOAD)
- в UisVapService добавлена реализация метода VapService findByVapName(....)
- в CustomInjectorUisDialogController:
 -- изменено поведение в customVapUisDialogModel.doneProperty().addListener(....):
  -> добавлено переключение на пользовательский список ВАП, если открыт список встроенных ВАП
  -> после удаления производится принудительное снятие выделения с элементов списка чтобы принудить пользователя явно подтвердить выбор ВАП и сохранить выбор
 -- исправлена ошибка очистки надписей с данными для второй катушки - пытались очистить названия надписей вместо значений; добавлены переменные для надписей самих значений
 -- в setDelete() и setEdit() добавлена установка выделения на ВАП, привязанный к этому инжектору
 -- извещение об окончании операций с инжектором вынесено в метод complete(Model injector)
- в CustomVapUisDialogController:
 -- реализованы операции EDIT, COPY
 -- действия по настройке элементов окна и данных в этом окне вынесены в методы disableNodes(boolean disable) и setValues() сооответственно
 -- в операции delete() добавлена проверка на отсутствие привязки этого ВАП к инжекторам и запрет удаления с открытием поясняющей надписи слева от кнопок в форме
 -- определена начальная установка кнопки enableBoostToggleButton, иначе неверно отображалась надпись на ней до первого клика
 -- в create() исправлена запись значений bipPwmSpinner и bipWindowSpinner: если == 0, то пишем в БД null (раньше писали 0 => потом неправильно обрабатывался признак установки BIP-галки)
 -- извещение об окончании операций с ВАП вынесено в метод complete(InjectorUisVAP vap)
- в CSVSUpdater изменена запись в файл: если null (раньше было == 0) - пишем пустую строку, причина изменения строкой выше

- в VoltAmpereProfile добавлен @NamedEntityGraph(name = "CrInjectorsByVAP"....) и getInjectors() для получения списка инжекторов, к которым привязан этот ВАП
- в VoltAmpereProfileRepository добавлен VoltAmpereProfile findByProfileName (String vapName) с @EntityGraph(value = "CrInjectorsByVAP", type = EntityGraph.EntityGraphType.LOAD)

- в NewEditInjectorDialogController:
 -- добавлена проверка на полноту данных с выбросом окна в случае отсутствия обязательных (по аналогии с UIS)
 -- извещение об окончании операций с инжектором вынесено в метод complete(Model injector)

- в NewEditVOAPDialogController:
 -- в операции удаления добавлена проверка на отсутствие привязки этого ВАП к инжекторам через метод findByProfileName(...) из VoltAmpereProfileRepository
 -- при положительном результате удаление не производится и открывается поясняющая надпись слева от кнопок в форме.
 -- добавлена установка значения в negativeUSpinner при настройке окна, была забыла
 -- определена начальная установка кнопки enableBoostToggleButton, иначе неверно отображалась надпись на ней до первого клика
 -- исправлена логика обработки состояния кнопки enableBoostToggleButton при записи этого параметра в БД
 -- добавлена перезагрузка списка ВАП при редактировании по факту сохранения в БД для адекватного отображения изменившихся значений

- в CustomVapUisDialog.fxml добавлена поясняющая надпись для случая отказа от удаления.
- в NewEditVOAPDialog.fxml добавлена поясняющая надпись для случая отказа от удаления.

+++ src/main/java/fi/stardex/sisu/ui/controllers/cr/dialogs/NewEditVOAPDialogController.java
@@ -17,6 +17,7 @@ import static fi.stardex.sisu.util.SpinnerDefaults.*;
 
 public class NewEditVOAPDialogController {
 
+    @FXML private Label deletionError;
     @FXML private Label firstW_1Label;
     @FXML private Label batteryULabel;
     @FXML private Label boostI_1Label;
@@ -165,6 +166,8 @@ public class NewEditVOAPDialogController {
         coil2CheckBox.selectedProperty().addListener((observableValue, oldValue, newValue) -> activateCoil2Spinners(newValue));
 
         bindingI18N();
+        enableBoostToggleButton.setSelected(false);
+        enableBoostToggleButton.setText(boostU_disabled.get());
     }
 
     private void activateCoil2Spinners(boolean activate) {
@@ -190,7 +193,7 @@ public class NewEditVOAPDialogController {
                     firstWSpinner.getValue(),
                     secondISpinner.getValue(),
                     negativeUSpinner.getValue(),
-                    enableBoostToggleButton.isSelected(),
+                    !enableBoostToggleButton.isSelected(),
                     boostI2Spinner.getValue(),
                     firstI2Spinner.getValue(),
                     firstW2Spinner.getValue(),
@@ -217,7 +220,7 @@ public class NewEditVOAPDialogController {
                 firstWSpinner.getValue(),
                 secondISpinner.getValue(),
                 negativeUSpinner.getValue(),
-                enableBoostToggleButton.isSelected(),
+                !enableBoostToggleButton.isSelected(),
                 boostI2Spinner.getValue(),
                 firstI2Spinner.getValue(),
                 firstW2Spinner.getValue(),
@@ -231,6 +234,12 @@ public class NewEditVOAPDialogController {
     }
 
     private void deleteVOAP() {
+
+        if (!voltAmpereProfileRepository.findByProfileName(voapList.getSelectionModel().getSelectedItem().getProfileName()).getInjectors().isEmpty()) {
+            deletionError.setVisible(true);
+            return;
+        }
+        deletionError.setVisible(false);
         voltAmpereProfileRepository.deleteById(voapList.getSelectionModel().getSelectedItem().getProfileName());
         voapList.getItems().remove(voapList.getSelectionModel().getSelectedItem());
         stage.close();
@@ -243,7 +252,7 @@ public class NewEditVOAPDialogController {
         vapToUpdate.setBatteryU(batteryUSpinner.getValue());
         vapToUpdate.setBoostU(boostUSpinner.getValue());
         vapToUpdate.setBoostI(boostISpinner.getValue());
-        vapToUpdate.setBoostDisable(enableBoostToggleButton.isSelected());
+        vapToUpdate.setBoostDisable(!enableBoostToggleButton.isSelected());
         vapToUpdate.setFirstI(firstISpinner.getValue());
         vapToUpdate.setFirstW(firstWSpinner.getValue());
         vapToUpdate.setSecondI(secondISpinner.getValue());
@@ -257,18 +266,23 @@ public class NewEditVOAPDialogController {
             vapToUpdate.setSecondI2(secondI2Spinner.getValue());
         }
         voltAmpereProfileRepository.save(vapToUpdate);
+        voapList.getItems().setAll(voltAmpereProfileRepository.findByIsCustomAndInjectorType(true, currentVOAP.getInjectorType()));
+        voapList.getSelectionModel().select(currentVOAP);
+        voapList.scrollTo(currentVOAP);
         stage.close();
     }
 
     public void setNew() {
         currentState = State.NEW;
         enterNameTF.clear();
+        deletionError.setVisible(false);
         disableNodes(false);
         enterNameTF.setDisable(false);
     }
 
     public void setCopy() {
         currentState = State.COPY;
+        deletionError.setVisible(false);
         disableNodes(true);
         setValues();
         enterNameTF.setDisable(false);
@@ -277,6 +291,7 @@ public class NewEditVOAPDialogController {
     public void setDelete() {
         currentState = State.DELETE;
         enterNameTF.setText(voapList.getSelectionModel().getSelectedItem().getProfileName());
+        deletionError.setVisible(false);
         disableNodes(true);
         setValues();
         enterNameTF.setDisable(true);
@@ -285,6 +300,7 @@ public class NewEditVOAPDialogController {
     public void setEdit() {
         currentState = State.EDIT;
         enterNameTF.setText(voapList.getSelectionModel().getSelectedItem().getProfileName());
+        deletionError.setVisible(false);
         disableNodes(false);
         setValues();
         enterNameTF.setDisable(true);
@@ -300,10 +316,18 @@ public class NewEditVOAPDialogController {
         negativeUSpinner.setDisable(disable);
         enableBoostToggleButton.setDisable(disable);
         coil2CheckBox.setDisable(disable);
-        boostI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
-        firstI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
-        firstW2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
-        secondI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
+
+        if (currentState != State.NEW) {
+            boostI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
+            firstI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
+            firstW2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
+            secondI2Spinner.setDisable(disable || !currentVOAP.isDoubleCoil());
+        } else {
+            boostI2Spinner.setDisable(disable || !coil2CheckBox.isSelected());
+            firstI2Spinner.setDisable(disable || !coil2CheckBox.isSelected());
+            firstW2Spinner.setDisable(disable || !coil2CheckBox.isSelected());
+            secondI2Spinner.setDisable(disable || !coil2CheckBox.isSelected());
+        }
     }
 
     private void setValues() {
@@ -315,7 +339,8 @@ public class NewEditVOAPDialogController {
         firstISpinner.getValueFactory().setValue(currentVOAP.getFirstI());
         firstWSpinner.getValueFactory().setValue(currentVOAP.getFirstW());
         secondISpinner.getValueFactory().setValue(currentVOAP.getSecondI());
-        enableBoostToggleButton.setSelected(currentVOAP.getBoostDisable());
+        negativeUSpinner.getValueFactory().setValue(currentVOAP.getNegativeU());
+        enableBoostToggleButton.setSelected(!currentVOAP.getBoostDisable());
         coil2CheckBox.setSelected(currentVOAP.isDoubleCoil());
 
         if (currentVOAP.isDoubleCoil()) {
@@ -384,5 +409,6 @@ public class NewEditVOAPDialogController {
         secondI_2Label.textProperty().bind(i18N.createStringBinding("h4.voltage.label.I2"));
         saveBtn.textProperty().bind(i18N.createStringBinding("h4.delay.button.save"));
         cancelBtn.textProperty().bind(i18N.createStringBinding("voapProfile.button.cancel"));
+        deletionError.textProperty().bind(i18N.createStringBinding("voapProfile.label.deletionError"));
     }
 }
